Set pianod2 port to 4446 in production
Was originally thought that we might serve both the UI and the websocket
on the same port. But it turns out that the pianod2 web server is not quite
there yet, so serving with nginx/apache is preferred.
Therefore our websocket port has to be 4446 instead of the same as the
website.